fix vlados
