+++ TanyaTest/src/test/java/TestFromVlados/UiTest/SpotifyTests/AddForPlaylistTest.java
@@ -1,19 +1,34 @@
 package TestFromVlados.UiTest.SpotifyTests;
 
+import TanyaTestFoxtrot.pageObjects.TestInit;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyHomePage;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyPage;
 import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class AddForPlaylistTest extends SpotifyPage {
+public class AddForPlaylistTest extends TestInit {
+
+    @BeforeMethod
+    public void loginSpotify() {
+        SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
+        SpotifyPage spotifyPage = new SpotifyPage(driver);
+        spotifyPage.goSpofify();
+        spotifyHomePage.loginButton().click();
+        spotifyHomePage.emailField().sendKeys(spotifyPage.emailSpotify());
+        spotifyHomePage.passwordField().sendKeys(spotifyPage.passwordSpotify());
+        spotifyHomePage.signButton().click();
+        spotifyHomePage.closeNotification().click();
+    }
 
     @Test
     public void addTrackForPlaylist() {
         SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
+        SpotifyPage spotifyPage = new SpotifyPage(driver);
         spotifyHomePage.searchButton().click();
-        spotifyHomePage.searchField().sendKeys(searchTrackOFD());
+        spotifyHomePage.searchField().sendKeys(spotifyPage.searchTrackOFD());
         spotifyHomePage.openTrack().click();
-        actions().contextClick(spotifyHomePage.moveToTrack()).perform();
+        spotifyPage.actions().contextClick(spotifyHomePage.moveToTrack()).perform();
         spotifyHomePage.likedTrack().click();
         spotifyHomePage.goLikedPlaylist().click();
         Assert.assertTrue(spotifyHomePage.checkTrackInPlaylist().isDisplayed());
