+++ TanyaTest/src/test/java/TestFromVlados/UiTest/SpotifyTests/FilterEpisodesTest.java
@@ -1,16 +1,31 @@
 package TestFromVlados.UiTest.SpotifyTests;
 
+import TanyaTestFoxtrot.pageObjects.TestInit;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyHomePage;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyPage;
 import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class FilterEpisodesTest extends SpotifyPage {
+public class FilterEpisodesTest extends TestInit {
+
+    @BeforeMethod
+    public void loginSpotify() {
+        SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
+        SpotifyPage spotifyPage = new SpotifyPage(driver);
+        spotifyPage.goSpofify();
+        spotifyHomePage.loginButton().click();
+        spotifyHomePage.emailField().sendKeys(spotifyPage.emailSpotify());
+        spotifyHomePage.passwordField().sendKeys(spotifyPage.passwordSpotify());
+        spotifyHomePage.signButton().click();
+        spotifyHomePage.closeNotification().click();
+    }
     @Test
     public void buttonEpisodes() {
         SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
+        SpotifyPage spotifyPage = new SpotifyPage(driver);
         spotifyHomePage.searchButton().click();
-        spotifyHomePage.searchField().sendKeys(searchTrackOFD());
+        spotifyHomePage.searchField().sendKeys(spotifyPage.searchTrackOFD());
         spotifyHomePage.buttonEpisodes().click();
         Assert.assertTrue(spotifyHomePage.assertButtonEpisodes().isDisplayed());
     }
