+++ TanyaTest/src/test/java/TestFromVlados/UiTest/SpotifyTests/OpenSettingsTest.java
@@ -1,11 +1,26 @@
 package TestFromVlados.UiTest.SpotifyTests;
 
+import TanyaTestFoxtrot.pageObjects.TestInit;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyHomePage;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyPage;
 import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class OpenSettingsTest extends SpotifyPage {
+public class OpenSettingsTest extends TestInit {
+
+    @BeforeMethod
+    public void loginSpotify() {
+        SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
+        SpotifyPage spotifyPage = new SpotifyPage(driver);
+        spotifyPage.goSpofify();
+        spotifyHomePage.loginButton().click();
+        spotifyHomePage.emailField().sendKeys(spotifyPage.emailSpotify());
+        spotifyHomePage.passwordField().sendKeys(spotifyPage.passwordSpotify());
+        spotifyHomePage.signButton().click();
+        spotifyHomePage.closeNotification().click();
+    }
+
     @Test
     public void openSettings() {
         SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
