+++ TanyaTest/src/test/java/TestFromVlados/UiTest/SpotifyTests/PlayTrackTest.java
@@ -1,17 +1,30 @@
 package TestFromVlados.UiTest.SpotifyTests;
-п
+import TanyaTestFoxtrot.pageObjects.TestInit;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyHomePage;
 import TestFromVlados.UiTest.ForSpotifyTests.SpotifyPage;
 import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class PlayTrackTest extends SpotifyPage {
+public class PlayTrackTest extends TestInit {
 
+    @BeforeMethod
+    public void loginSpotify() {
+        SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
+        SpotifyPage spotifyPage = new SpotifyPage(driver);
+        spotifyPage.goSpofify();
+        spotifyHomePage.loginButton().click();
+        spotifyHomePage.emailField().sendKeys(spotifyPage.emailSpotify());
+        spotifyHomePage.passwordField().sendKeys(spotifyPage.passwordSpotify());
+        spotifyHomePage.signButton().click();
+        spotifyHomePage.closeNotification().click();
+    }
     @Test
     public void playMusik() {
         SpotifyHomePage spotifyHomePage = new SpotifyHomePage(driver);
+        SpotifyPage spotifyPage = new SpotifyPage(driver);
         spotifyHomePage.searchButton().click();
-        spotifyHomePage.searchField().sendKeys(searchTrackOFD());
+        spotifyHomePage.searchField().sendKeys(spotifyPage.searchTrackOFD());
         spotifyHomePage.openTrack().click();
         spotifyHomePage.playTrack().click();
         Assert.assertTrue(spotifyHomePage.assertPause().isDisplayed());
